Add more permissive permissions to the /opt/Ankama folder to prevent failing updates
The Ankama Launcher apparently needs more permissive permissions to perform update as of now.
Performing update with a classic `755` perms resulted in:
> Error: EACCES: permission denied, unlink '/opt/Ankama/Ankama-Launcher-x86_64.AppImage'

While `777` perms are usually to avoid, settings this kind of permissives permissions on the `/opt/Ankama` directory within the container fixes the issue and shouldn't cause any security concern (since this directory is located within the container itself and only contains the appimage).