Rebuild to address the xz backdoor/security issue + switch to Arch's DockerHub repository
About the xz matter:
https://www.openwall.com/lists/oss-security/2024/03/29/4
https://archlinux.org/news/the-xz-package-has-been-backdoored/
https://security.archlinux.org/ASA-202403-1

Arch Linux container images are built daily on Arch's DockerHub repository (compared to being built weekly on the official DockerHub library).
Switching to a daily updated source allows to act faster on such eventual issue (as, for instance, the "latest" Arch image on the official DockerHub library currently is 14 days old, and do not contain the fix yet).